refactoring data generation code into helper class
