+++ tests/dao_tests/test_ehr_dao_test.py
@@ -46,7 +46,7 @@ class EhrReceiptDaoTest(BaseTestCase):
         return org
 
     def _make_participant(self, org, int_id):
-        participant = self._participant_with_defaults(participantId=int_id, biobankId=int_id)
+        participant = self.data_generator._participant_with_defaults(participantId=int_id, biobankId=int_id)
         participant.hpoId = org.hpoId
         participant.organizationId = org.organizationId
         self.participant_dao.insert(participant)
